+++ .babelrc
@@ -5,7 +5,8 @@
       {
         "modules": false,
         "targets": {
-          "node": "current"
+          "browsers": "> 1%",
+          "uglify": true
         },
         "useBuiltIns": true
       }
@@ -14,6 +15,7 @@
   ],
   "plugins": [
     "syntax-dynamic-import",
+    "transform-runtime",
     "transform-object-rest-spread",
     [
       "transform-class-properties",
