+++ package.json
@@ -3,6 +3,7 @@
   "private": true,
   "dependencies": {
     "@rails/webpacker": "^3.3.1",
+    "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-react": "^6.24.1",
     "bootstrap-sass": "^3.3.7",
     "cheerio": "^1.0.0-rc.2",
